Updated turtlerace.py
Added a for loop to allow the user to choose how many times the turtles race. Also added a way to keep score of how many times each turtle wins, you do have to create a file named: "TurtleRaceWinners.txt" in the same directory as the "turtlerace.py" file. And writes winner color of turtle.